Address PR feedback: clean up types, graceful cancellation, extract class
- Use IExtensionWrapper directly (with existing using) in
  ExtensionGalleryViewModel instead of fully-qualified name
- Replace ct.ThrowIfCancellationRequested() with graceful early return
  in BuiltInExtensionService.LoadProvidersAsync
- Extract ExtensionStartResult into its own file
- Restore 'Go back to the main page, but keep it open' comment
