+++ src/modules/cmdpal/Microsoft.CmdPal.UI.ViewModels/Services/WinRTExtensionService.cs
@@ -555,43 +555,6 @@ public partial class WinRTExtensionService : IExtensionService, IDisposable
             _disposedValue = true;
         }
     }
-
-    private sealed class ExtensionStartResult
-    {
-        public IExtensionWrapper Extension { get; }
-
-        public CommandProviderWrapper? Wrapper { get; private init; }
-
-        public Task? PendingStartTask { get; private init; }
-
-        public Stopwatch? Stopwatch { get; private init; }
-
-        [System.Diagnostics.CodeAnalysis.MemberNotNullWhen(true, nameof(Wrapper))]
-        public bool IsStarted => Wrapper is not null;
-
-        [System.Diagnostics.CodeAnalysis.MemberNotNullWhen(true, nameof(PendingStartTask), nameof(Stopwatch))]
-        public bool IsTimedOut => PendingStartTask is not null;
-
-        private ExtensionStartResult(IExtensionWrapper extension)
-        {
-            Extension = extension;
-        }
-
-        public static ExtensionStartResult Started(IExtensionWrapper extension, CommandProviderWrapper wrapper)
-        {
-            return new ExtensionStartResult(extension) { Wrapper = wrapper };
-        }
-
-        public static ExtensionStartResult TimedOut(IExtensionWrapper extension, Task pendingStartTask, Stopwatch sw)
-        {
-            return new ExtensionStartResult(extension) { PendingStartTask = pendingStartTask, Stopwatch = sw };
-        }
-
-        public static ExtensionStartResult Failed(IExtensionWrapper extension)
-        {
-            return new ExtensionStartResult(extension);
-        }
-    }
 }
 
 internal record struct IsExtensionResult(bool IsExtension, AppExtension? Extension)
